tests, namespaces updated
