+++ sdk/digitaltwins/Azure.DigitalTwins.Core/src/DigitalTwinsClient.cs
@@ -10,7 +10,8 @@ using System.Threading.Tasks;
 using Azure.Core;
 using Azure.Core.Pipeline;
 using Azure.Core.Serialization;
-using Azure.DigitalTwins.Core.QueryBuilder;
+using Fluent = Azure.DigitalTwins.Core.QueryBuilder.Fluent;
+using Linq = Azure.DigitalTwins.Core.QueryBuilder.Linq;
 using static Azure.DigitalTwins.Core.StreamHelper;
 
 namespace Azure.DigitalTwins.Core
@@ -2148,7 +2149,7 @@ namespace Azure.DigitalTwins.Core
         /// <param name="cancellationToken">The cancellation token.</param>
         /// <returns>The pageable list <see cref="AsyncPageable{T}"/> of query results.</returns>
         /// <typeparam name="T">The type to deserialize the result to.</typeparam>
-        public virtual AsyncPageable<T> QueryAsync<T>(DigitalTwinsQueryBuilderMethodDriven adtQueryBuilder, CancellationToken cancellationToken = default)
+        public virtual AsyncPageable<T> QueryAsync<T>(Fluent.DigitalTwinsQueryBuilder adtQueryBuilder, CancellationToken cancellationToken = default)
         {
             using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(DigitalTwinsClient)}.{nameof(Query)}");
             scope.Start();
@@ -2173,7 +2174,57 @@ namespace Azure.DigitalTwins.Core
         /// <param name="cancellationToken">The cancellation token.</param>
         /// <returns>The pageable list <see cref="Pageable{T}"/> of query results.</returns>
         /// <typeparam name="T">The type to deserialize the result to.</typeparam>
-        public virtual Pageable<T> Query<T>(DigitalTwinsQueryBuilderMethodDriven adtQueryBuilder, CancellationToken cancellationToken = default)
+        public virtual Pageable<T> Query<T>(Fluent.DigitalTwinsQueryBuilder adtQueryBuilder, CancellationToken cancellationToken = default)
+        {
+            using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(DigitalTwinsClient)}.{nameof(Query)}");
+            scope.Start();
+
+            Argument.AssertNotNull(adtQueryBuilder, nameof(adtQueryBuilder));
+
+            try
+            {
+                return Query<T>(adtQueryBuilder.GetQueryText(), cancellationToken);
+            }
+            catch (Exception ex)
+            {
+                scope.Failed(ex);
+                throw;
+            }
+        }
+
+        /// <summary>
+        /// Queries for digital twins by iterating through a collection asynchronously.
+        /// </summary>
+        /// <param name="adtQueryBuilder"> Query object made with the ADT query builder. </param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The pageable list <see cref="AsyncPageable{T}"/> of query results.</returns>
+        /// <typeparam name="T">The type to deserialize the result to.</typeparam>
+        public virtual AsyncPageable<T> QueryAsync<T>(Linq.DigitalTwinsQueryBuilder adtQueryBuilder, CancellationToken cancellationToken = default)
+        {
+            using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(DigitalTwinsClient)}.{nameof(Query)}");
+            scope.Start();
+
+            Argument.AssertNotNull(adtQueryBuilder, nameof(adtQueryBuilder));
+
+            try
+            {
+                return QueryAsync<T>(adtQueryBuilder.GetQueryText(), cancellationToken);
+            }
+            catch (Exception ex)
+            {
+                scope.Failed(ex);
+                throw;
+            }
+        }
+
+        /// <summary>
+        /// Queries for digital twins by iterating through a collection synchronously.
+        /// </summary>
+        /// <param name="adtQueryBuilder"> Query object made with the ADT query builder. </param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The pageable list <see cref="Pageable{T}"/> of query results.</returns>
+        /// <typeparam name="T">The type to deserialize the result to.</typeparam>
+        public virtual Pageable<T> Query<T>(Linq.DigitalTwinsQueryBuilder adtQueryBuilder, CancellationToken cancellationToken = default)
         {
             using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(DigitalTwinsClient)}.{nameof(Query)}");
             scope.Start();
