+++ sdk/digitaltwins/Azure.DigitalTwins.Core/tests/DigitalTwinsQueryBuilderTests/ConferenceRoom.cs
